First upload
A simple game of rock, paper, scissors. User will input their choice between the 3 and computer will randomly choose between them as well. Has a few lines for UX. Will clean up UX in the future to make it a bit easier to read and interact with. Will be adding an exit feature as well just to do it. May create GUI for this as well for fun.